Refactoring the Xtend IDE rename refactoring test cases.
- Eliminate duplicated code by introducing the
AbstractXtendRenameRefactoringTest abstract base class.
- Add missing/update existing copyright headers.
