+++ org.eclipse.xtend.ide.tests/xtend-gen/org/eclipse/xtend/ide/tests/refactoring/RenameResourceTest.java
@@ -1,45 +1,25 @@
+/**
+ * Copyright (c) 2012, 2018 itemis AG (http://www.itemis.eu) and others.
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License v1.0
+ * which accompanies this distribution, and is available at
+ * http://www.eclipse.org/legal/epl-v10.html
+ */
 package org.eclipse.xtend.ide.tests.refactoring;
 
-import com.google.inject.Inject;
-import org.eclipse.core.resources.IFile;
-import org.eclipse.core.resources.IProject;
-import org.eclipse.core.resources.IResource;
-import org.eclipse.core.resources.IWorkspace;
-import org.eclipse.core.resources.IWorkspaceRunnable;
-import org.eclipse.core.runtime.IProgressMonitor;
-import org.eclipse.core.runtime.NullProgressMonitor;
-import org.eclipse.core.runtime.Path;
-import org.eclipse.ltk.core.refactoring.Change;
-import org.eclipse.ltk.core.refactoring.participants.RenameRefactoring;
-import org.eclipse.ltk.internal.core.refactoring.resource.RenameResourceProcessor;
-import org.eclipse.xtend.ide.tests.AbstractXtendUITestCase;
-import org.eclipse.xtend.ide.tests.WorkbenchTestHelper;
-import org.eclipse.xtend.ide.tests.refactoring.FileAsserts;
+import org.eclipse.xtend.ide.tests.refactoring.AbstractXtendRenameRefactoringTest;
 import org.eclipse.xtext.xbase.lib.Exceptions;
-import org.eclipse.xtext.xbase.lib.Extension;
-import org.junit.Assert;
 import org.junit.Test;
 
 @SuppressWarnings("all")
-public class RenameResourceTest extends AbstractXtendUITestCase {
-  @Inject
-  @Extension
-  private WorkbenchTestHelper _workbenchTestHelper;
-  
-  @Inject
-  private IWorkspace workspace;
-  
-  @Inject
-  @Extension
-  private FileAsserts _fileAsserts;
-  
+public class RenameResourceTest extends AbstractXtendRenameRefactoringTest {
   @Test
   public void testRenameClassWithSameName() {
     try {
       try {
-        this._fileAsserts.assertFileContains(this.renameTo(this._workbenchTestHelper.createFile("Foo", "class Foo {}"), "Bar.xtend"), "class Bar {}");
+        this.fileAsserts.assertFileContains(this.renameTo(this.testHelper.createFile("Foo", "class Foo {}"), "Bar.xtend"), "class Bar {}");
       } finally {
-        this._workbenchTestHelper.getFile("Bar.xtend").delete(true, null);
+        this.testHelper.getFile("Bar.xtend").delete(true, null);
       }
     } catch (Throwable _e) {
       throw Exceptions.sneakyThrow(_e);
@@ -50,9 +30,9 @@ public class RenameResourceTest extends AbstractXtendUITestCase {
   public void testDontRenameClassWithDifferenName() {
     try {
       try {
-        this._fileAsserts.assertFileContains(this.renameTo(this._workbenchTestHelper.createFile("Foo", "class FooBar {}"), "Bar.xtend"), "class FooBar {}");
+        this.fileAsserts.assertFileContains(this.renameTo(this.testHelper.createFile("Foo", "class FooBar {}"), "Bar.xtend"), "class FooBar {}");
       } finally {
-        this._workbenchTestHelper.getFile("Bar.xtend").delete(true, null);
+        this.testHelper.getFile("Bar.xtend").delete(true, null);
       }
     } catch (Throwable _e) {
       throw Exceptions.sneakyThrow(_e);
@@ -63,39 +43,10 @@ public class RenameResourceTest extends AbstractXtendUITestCase {
   public void testGuardMissingFileExtension() {
     try {
       try {
-        this._fileAsserts.assertFileContains(this.renameTo(this._workbenchTestHelper.createFile("Foo", "class Foo {}"), "Bar"), "class Foo {}");
+        this.fileAsserts.assertFileContains(this.renameTo(this.testHelper.createFile("Foo", "class Foo {}"), "Bar"), "class Foo {}");
       } finally {
-        this._workbenchTestHelper.getFile("Bar").delete(true, null);
+        this.testHelper.getFile("Bar").delete(true, null);
       }
-    } catch (Throwable _e) {
-      throw Exceptions.sneakyThrow(_e);
-    }
-  }
-  
-  protected IFile renameTo(final IFile file, final String newFileName) {
-    try {
-      IFile _xblockexpression = null;
-      {
-        final RenameResourceProcessor renameResourceProcessor = new RenameResourceProcessor(file);
-        renameResourceProcessor.setNewResourceName(newFileName);
-        final RenameRefactoring renameRefactoring = new RenameRefactoring(renameResourceProcessor);
-        NullProgressMonitor _nullProgressMonitor = new NullProgressMonitor();
-        renameRefactoring.checkAllConditions(_nullProgressMonitor);
-        NullProgressMonitor _nullProgressMonitor_1 = new NullProgressMonitor();
-        final Change change = renameRefactoring.createChange(_nullProgressMonitor_1);
-        final IWorkspaceRunnable _function = (IProgressMonitor it) -> {
-          change.perform(it);
-        };
-        NullProgressMonitor _nullProgressMonitor_2 = new NullProgressMonitor();
-        this.workspace.run(_function, _nullProgressMonitor_2);
-        IProject _project = this._workbenchTestHelper.getProject();
-        Path _path = new Path(("src/" + newFileName));
-        final IResource newFile = _project.findMember(_path);
-        Assert.assertTrue(newFile.exists());
-        Assert.assertTrue((newFile instanceof IFile));
-        _xblockexpression = ((IFile) newFile);
-      }
-      return _xblockexpression;
     } catch (Throwable _e) {
       throw Exceptions.sneakyThrow(_e);
     }
