+++ org.eclipse.xtend.ide.tests/xtend-gen/org/eclipse/xtend/ide/tests/refactoring/XImportSectionUpdateOnRenameTest.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2014 itemis AG (http://www.itemis.eu) and others.
+ * Copyright (c) 2014, 2018 itemis AG (http://www.itemis.eu) and others.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
  * which accompanies this distribution, and is available at
@@ -7,24 +7,11 @@
  */
 package org.eclipse.xtend.ide.tests.refactoring;
 
-import com.google.inject.Inject;
 import org.eclipse.core.resources.IFile;
 import org.eclipse.core.runtime.NullProgressMonitor;
-import org.eclipse.jdt.core.IField;
-import org.eclipse.jdt.core.IMethod;
-import org.eclipse.jdt.core.IType;
-import org.eclipse.jdt.core.JavaCore;
-import org.eclipse.jdt.ui.refactoring.RenameSupport;
-import org.eclipse.ui.IWorkbench;
-import org.eclipse.xtend.ide.tests.AbstractXtendUITestCase;
-import org.eclipse.xtend.ide.tests.WorkbenchTestHelper;
-import org.eclipse.xtend.ide.tests.refactoring.FileAsserts;
+import org.eclipse.xtend.ide.tests.refactoring.AbstractXtendRenameRefactoringTest;
 import org.eclipse.xtend2.lib.StringConcatenation;
-import org.eclipse.xtext.common.types.ui.refactoring.participant.CompositeRefactoringProcessor;
-import org.eclipse.xtext.ui.refactoring.ui.SyncUtil;
 import org.eclipse.xtext.xbase.lib.Exceptions;
-import org.eclipse.xtext.xbase.lib.Extension;
-import org.junit.Assert;
 import org.junit.Ignore;
 import org.junit.Test;
 
@@ -32,30 +19,7 @@ import org.junit.Test;
  * @author Anton Kosyakov - Initial contribution and API
  */
 @SuppressWarnings("all")
-public class XImportSectionUpdateOnRenameTest extends AbstractXtendUITestCase {
-  @Inject
-  private SyncUtil syncUtil;
-  
-  @Inject
-  private IWorkbench workbench;
-  
-  @Inject
-  @Extension
-  private FileAsserts _fileAsserts;
-  
-  @Inject
-  @Extension
-  private WorkbenchTestHelper testHelper;
-  
-  @Inject
-  private CompositeRefactoringProcessor.Access compositeRefactoringProcessorAccess;
-  
-  @Override
-  public void tearDown() throws Exception {
-    this.testHelper.tearDown();
-    super.tearDown();
-  }
-  
+public class XImportSectionUpdateOnRenameTest extends AbstractXtendRenameRefactoringTest {
   @Test
   public void renameUnusedType() {
     try {
@@ -87,7 +51,7 @@ public class XImportSectionUpdateOnRenameTest extends AbstractXtendUITestCase {
         _builder_2.newLine();
         _builder_2.append("class Bar {}");
         _builder_2.newLine();
-        this._fileAsserts.assertFileContains(xtendFile, _builder_2.toString());
+        this.fileAsserts.assertFileContains(xtendFile, _builder_2.toString());
       } finally {
         IFile _file = this.testHelper.getProject().getFile("src/foo/NewFoo.java");
         NullProgressMonitor _nullProgressMonitor = new NullProgressMonitor();
@@ -130,7 +94,7 @@ public class XImportSectionUpdateOnRenameTest extends AbstractXtendUITestCase {
         _builder_2.newLine();
         _builder_2.append("class Bar {}");
         _builder_2.newLine();
-        this._fileAsserts.assertFileContains(xtendFile, _builder_2.toString());
+        this.fileAsserts.assertFileContains(xtendFile, _builder_2.toString());
       } finally {
         IFile _file = this.testHelper.getProject().getFile("src/foo/NewFoo.java");
         NullProgressMonitor _nullProgressMonitor = new NullProgressMonitor();
@@ -179,7 +143,7 @@ public class XImportSectionUpdateOnRenameTest extends AbstractXtendUITestCase {
         _builder_2.newLine();
         _builder_2.append("class Bar {}");
         _builder_2.newLine();
-        this._fileAsserts.assertFileContains(xtendFile, _builder_2.toString());
+        this.fileAsserts.assertFileContains(xtendFile, _builder_2.toString());
       } finally {
         IFile _file = this.testHelper.getProject().getFile("src/foo/NewFoo.java");
         NullProgressMonitor _nullProgressMonitor = new NullProgressMonitor();
@@ -228,7 +192,7 @@ public class XImportSectionUpdateOnRenameTest extends AbstractXtendUITestCase {
         _builder_2.newLine();
         _builder_2.append("class Bar {}");
         _builder_2.newLine();
-        this._fileAsserts.assertFileContains(xtendFile, _builder_2.toString());
+        this.fileAsserts.assertFileContains(xtendFile, _builder_2.toString());
       } finally {
         IFile _file = this.testHelper.getProject().getFile("src/foo/NewFoo.java");
         NullProgressMonitor _nullProgressMonitor = new NullProgressMonitor();
@@ -277,7 +241,7 @@ public class XImportSectionUpdateOnRenameTest extends AbstractXtendUITestCase {
         _builder_2.newLine();
         _builder_2.append("class Bar {}");
         _builder_2.newLine();
-        this._fileAsserts.assertFileContains(xtendFile, _builder_2.toString());
+        this.fileAsserts.assertFileContains(xtendFile, _builder_2.toString());
       } finally {
         IFile _file = this.testHelper.getProject().getFile("src/foo/NewFoo.java");
         NullProgressMonitor _nullProgressMonitor = new NullProgressMonitor();
@@ -326,7 +290,7 @@ public class XImportSectionUpdateOnRenameTest extends AbstractXtendUITestCase {
         _builder_2.newLine();
         _builder_2.append("class Bar {}");
         _builder_2.newLine();
-        this._fileAsserts.assertFileContains(xtendFile, _builder_2.toString());
+        this.fileAsserts.assertFileContains(xtendFile, _builder_2.toString());
       } finally {
         IFile _file = this.testHelper.getProject().getFile("src/foo/NewFoo.java");
         NullProgressMonitor _nullProgressMonitor = new NullProgressMonitor();
@@ -374,7 +338,7 @@ public class XImportSectionUpdateOnRenameTest extends AbstractXtendUITestCase {
       _builder_2.newLine();
       _builder_2.append("class Bar {}");
       _builder_2.newLine();
-      this._fileAsserts.assertFileContains(xtendFile, _builder_2.toString());
+      this.fileAsserts.assertFileContains(xtendFile, _builder_2.toString());
     } catch (Throwable _e) {
       throw Exceptions.sneakyThrow(_e);
     }
@@ -417,7 +381,7 @@ public class XImportSectionUpdateOnRenameTest extends AbstractXtendUITestCase {
       _builder_2.newLine();
       _builder_2.append("class Bar {}");
       _builder_2.newLine();
-      this._fileAsserts.assertFileContains(xtendFile, _builder_2.toString());
+      this.fileAsserts.assertFileContains(xtendFile, _builder_2.toString());
     } catch (Throwable _e) {
       throw Exceptions.sneakyThrow(_e);
     }
@@ -454,7 +418,7 @@ public class XImportSectionUpdateOnRenameTest extends AbstractXtendUITestCase {
         _builder_2.newLine();
         _builder_2.append("class Bar extends NewFoo {}");
         _builder_2.newLine();
-        this._fileAsserts.assertFileContains(xtendFile, _builder_2.toString());
+        this.fileAsserts.assertFileContains(xtendFile, _builder_2.toString());
       } finally {
         IFile _file = this.testHelper.getProject().getFile("src/foo/NewFoo.java");
         NullProgressMonitor _nullProgressMonitor = new NullProgressMonitor();
@@ -504,7 +468,7 @@ public class XImportSectionUpdateOnRenameTest extends AbstractXtendUITestCase {
         _builder_2.newLine();
         _builder_2.append("class Bar extends Inner {}");
         _builder_2.newLine();
-        this._fileAsserts.assertFileContains(xtendFile, _builder_2.toString());
+        this.fileAsserts.assertFileContains(xtendFile, _builder_2.toString());
       } finally {
         IFile _file = this.testHelper.getProject().getFile("src/foo/NewFoo.java");
         NullProgressMonitor _nullProgressMonitor = new NullProgressMonitor();
@@ -579,7 +543,7 @@ public class XImportSectionUpdateOnRenameTest extends AbstractXtendUITestCase {
         _builder_2.newLine();
         _builder_2.append("}");
         _builder_2.newLine();
-        this._fileAsserts.assertFileContains(xtendFile, _builder_2.toString());
+        this.fileAsserts.assertFileContains(xtendFile, _builder_2.toString());
       } finally {
         IFile _file = this.testHelper.getProject().getFile("src/foo/NewFoo.java");
         NullProgressMonitor _nullProgressMonitor = new NullProgressMonitor();
@@ -654,7 +618,7 @@ public class XImportSectionUpdateOnRenameTest extends AbstractXtendUITestCase {
         _builder_2.newLine();
         _builder_2.append("}");
         _builder_2.newLine();
-        this._fileAsserts.assertFileContains(xtendFile, _builder_2.toString());
+        this.fileAsserts.assertFileContains(xtendFile, _builder_2.toString());
       } finally {
         IFile _file = this.testHelper.getProject().getFile("src/foo/NewFoo.java");
         NullProgressMonitor _nullProgressMonitor = new NullProgressMonitor();
@@ -729,7 +693,7 @@ public class XImportSectionUpdateOnRenameTest extends AbstractXtendUITestCase {
         _builder_2.newLine();
         _builder_2.append("}");
         _builder_2.newLine();
-        this._fileAsserts.assertFileContains(xtendFile, _builder_2.toString());
+        this.fileAsserts.assertFileContains(xtendFile, _builder_2.toString());
       } finally {
         IFile _file = this.testHelper.getProject().getFile("src/foo/NewFoo.java");
         NullProgressMonitor _nullProgressMonitor = new NullProgressMonitor();
@@ -804,7 +768,7 @@ public class XImportSectionUpdateOnRenameTest extends AbstractXtendUITestCase {
         _builder_2.newLine();
         _builder_2.append("}");
         _builder_2.newLine();
-        this._fileAsserts.assertFileContains(xtendFile, _builder_2.toString());
+        this.fileAsserts.assertFileContains(xtendFile, _builder_2.toString());
       } finally {
         IFile _file = this.testHelper.getProject().getFile("src/foo/NewFoo.java");
         NullProgressMonitor _nullProgressMonitor = new NullProgressMonitor();
@@ -882,7 +846,7 @@ public class XImportSectionUpdateOnRenameTest extends AbstractXtendUITestCase {
         _builder_3.newLine();
         _builder_3.append("}");
         _builder_3.newLine();
-        this._fileAsserts.assertFileContains(xtendFile, _builder_3.toString());
+        this.fileAsserts.assertFileContains(xtendFile, _builder_3.toString());
       } finally {
         IFile _file = this.testHelper.getProject().getFile("src/a/B.java");
         NullProgressMonitor _nullProgressMonitor = new NullProgressMonitor();
@@ -960,7 +924,7 @@ public class XImportSectionUpdateOnRenameTest extends AbstractXtendUITestCase {
         _builder_3.newLine();
         _builder_3.append("}");
         _builder_3.newLine();
-        this._fileAsserts.assertFileContains(xtendFile, _builder_3.toString());
+        this.fileAsserts.assertFileContains(xtendFile, _builder_3.toString());
       } finally {
         IFile _file = this.testHelper.getProject().getFile("src/a/B.java");
         NullProgressMonitor _nullProgressMonitor = new NullProgressMonitor();
@@ -1038,7 +1002,7 @@ public class XImportSectionUpdateOnRenameTest extends AbstractXtendUITestCase {
         _builder_3.newLine();
         _builder_3.append("}");
         _builder_3.newLine();
-        this._fileAsserts.assertFileContains(xtendFile, _builder_3.toString());
+        this.fileAsserts.assertFileContains(xtendFile, _builder_3.toString());
       } finally {
         IFile _file = this.testHelper.getProject().getFile("src/a/B.java");
         NullProgressMonitor _nullProgressMonitor = new NullProgressMonitor();
@@ -1116,7 +1080,7 @@ public class XImportSectionUpdateOnRenameTest extends AbstractXtendUITestCase {
         _builder_3.newLine();
         _builder_3.append("}");
         _builder_3.newLine();
-        this._fileAsserts.assertFileContains(xtendFile, _builder_3.toString());
+        this.fileAsserts.assertFileContains(xtendFile, _builder_3.toString());
       } finally {
         IFile _file = this.testHelper.getProject().getFile("src/a/B.java");
         NullProgressMonitor _nullProgressMonitor = new NullProgressMonitor();
@@ -1190,7 +1154,7 @@ public class XImportSectionUpdateOnRenameTest extends AbstractXtendUITestCase {
       _builder_2.newLine();
       _builder_2.append("}");
       _builder_2.newLine();
-      this._fileAsserts.assertFileContains(xtendFile, _builder_2.toString());
+      this.fileAsserts.assertFileContains(xtendFile, _builder_2.toString());
     } catch (Throwable _e) {
       throw Exceptions.sneakyThrow(_e);
     }
@@ -1259,42 +1223,9 @@ public class XImportSectionUpdateOnRenameTest extends AbstractXtendUITestCase {
       _builder_2.newLine();
       _builder_2.append("}");
       _builder_2.newLine();
-      this._fileAsserts.assertFileContains(xtendFile, _builder_2.toString());
+      this.fileAsserts.assertFileContains(xtendFile, _builder_2.toString());
     } catch (Throwable _e) {
       throw Exceptions.sneakyThrow(_e);
     }
-  }
-  
-  public IType findJavaType(final String typeName) throws Exception {
-    IType _xblockexpression = null;
-    {
-      this.syncUtil.totalSync(false);
-      _xblockexpression = JavaCore.create(this.testHelper.getProject()).findType(typeName);
-    }
-    return _xblockexpression;
-  }
-  
-  public void renameJavaElement(final IType javaElement, final String newName) throws Exception {
-    this.syncUtil.totalSync(false);
-    final RenameSupport renameSupport = RenameSupport.create(javaElement, newName, RenameSupport.UPDATE_REFERENCES);
-    renameSupport.perform(this.workbench.getActiveWorkbenchWindow().getShell(), this.workbench.getActiveWorkbenchWindow());
-    this.syncUtil.totalSync(false);
-    Assert.assertTrue(this.compositeRefactoringProcessorAccess.isDisposed());
-  }
-  
-  public void renameJavaElement(final IMethod javaElement, final String newName) throws Exception {
-    this.syncUtil.totalSync(false);
-    final RenameSupport renameSupport = RenameSupport.create(javaElement, newName, RenameSupport.UPDATE_REFERENCES);
-    renameSupport.perform(this.workbench.getActiveWorkbenchWindow().getShell(), this.workbench.getActiveWorkbenchWindow());
-    this.syncUtil.totalSync(false);
-    Assert.assertTrue(this.compositeRefactoringProcessorAccess.isDisposed());
-  }
-  
-  public void renameJavaElement(final IField javaElement, final String newName) throws Exception {
-    this.syncUtil.totalSync(false);
-    final RenameSupport renameSupport = RenameSupport.create(javaElement, newName, RenameSupport.UPDATE_REFERENCES);
-    renameSupport.perform(this.workbench.getActiveWorkbenchWindow().getShell(), this.workbench.getActiveWorkbenchWindow());
-    this.syncUtil.totalSync(false);
-    Assert.assertTrue(this.compositeRefactoringProcessorAccess.isDisposed());
   }
 }
